stop relying on From<u64> for Bson

+++ src/operation/count/test.rs
@@ -81,7 +81,7 @@ async fn handle_success() {
     let count_op = Count::empty();
 
     let n = 26;
-    let response = doc! { "ok": 1.0, "n": n };
+    let response = doc! { "ok": 1.0, "n": n as i32 };
 
     let actual_values = handle_response_test(&count_op, response).unwrap();
     assert_eq!(actual_values, n);
@@ -98,7 +98,7 @@ async fn handle_success_agg() {
         "cursor": {
             "id": 0,
             "ns": "a.b",
-            "firstBatch": [ { "n": n } ]
+            "firstBatch": [ { "n": n as i32 } ]
         }
     };
